feat: edited homepage, changed foulder names
