+++ data/db.json
@@ -20,6 +20,20 @@
       "category": "outdoor",
       "image": "https://images.unsplash.com/photo-1529313779164-31ef6a34524f?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1470&q=80",
       "id": 3
+    },
+    {
+      "title": "sadasd",
+      "details": "sadasdasd",
+      "image": "https://images.unsplash.com/photo-1529313779164-31ef6a34524f?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1470&q=80",
+      "category": "cars",
+      "id": 4
+    },
+    {
+      "title": "sadasd",
+      "details": "sadasdasd",
+      "image": "https://images.unsplash.com/photo-1529313779164-31ef6a34524f?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1470&q=80",
+      "category": "cars",
+      "id": 5
     }
   ]
 }
