+++ src/pages/index.jsx
@@ -1,8 +1,8 @@
 import Head from "next/head";
 import React from "react";
-import Button from "../atoms/button";
-import useGet from "../ions/hooks/fetch/get";
+import Button from "@mui/material/Button";
 import Layout from "../organisms/layout";
+import Link from "next/link";
 
 // fontsource/roboto
 import "@fontsource/roboto/300.css";
@@ -11,8 +11,6 @@ import "@fontsource/roboto/500.css";
 import "@fontsource/roboto/700.css";
 
 const Page = () => {
-	const { data, loading, error } = useGet("/api/hello");
-
 	return (
 		<Layout>
 			<Head>
@@ -20,14 +18,11 @@ const Page = () => {
 				<meta key="description" name="description" content="This is my project" />
 			</Head>
 			<h1>Home</h1>
-			{loading && <div>Loading...</div>}
-			{error && <div>{error.message}</div>}
-			{data && (
-				<pre>
-					<code>{JSON.stringify(data, null, 4)}</code>
-				</pre>
-			)}
-			<Button>Click me</Button>
+			<div style={{ display: "flex", justifyContent: "center" }}>
+				<Button variant="outlined">
+					<Link href="/form">Add a Item</Link>
+				</Button>
+			</div>
 		</Layout>
 	);
 };
